refactor(core): 迁移 catalog 管理路由到 admin 域
