+++ noj-core/drizzle/meta/_journal.json
@@ -540,6 +540,13 @@
       "when": 1788826201876,
       "tag": "0076_spicy_phalanx",
       "breakpoints": true
+    },
+    {
+      "idx": 77,
+      "version": "7",
+      "when": 1788828998366,
+      "tag": "0077_worthless_hemingway",
+      "breakpoints": true
     }
   ]
 }
