+++ noj-core/src/domains/admin/index.ts
@@ -3,18 +3,19 @@
  *
  * 提供（挂载前缀 /api/v1/admin，见 app.ts）：
  * - /identity/users、/identity/roles、/identity/permissions、/identity/blacklist
- * - /problems、/submissions、/contests、/judge-images、
- *   /dashboard/stats、/settings、/audit-logs、/announcements、/trainings、/llm/...
+ * - /catalog/problems、/catalog/trainings
+ * - /submissions、/contests、/judge-images、
+ *   /dashboard/stats、/settings、/audit-logs、/announcements、/llm/...
  *
  * 组级守卫：所有 admin 端点均需认证 + 管理员权限，在此统一挂载。
- * 例外：公告与题单管理端点已抽至独立 router，使用各自的细粒度权限。
+ * 例外：公告与 catalog 子域已抽至独立 router，使用各自的细粒度权限。
  * 此处必须跳过对应路径，否则组级通配 use 会先于独立 router 拦截请求。
  */
 import { Hono } from "hono";
 import type { AuthEnv } from "../identity/index.ts";
 import { adminMiddleware, authMiddleware } from "../identity/index.ts";
 import identityAdminRouter from "./routes/identity.ts";
-import { catalogAdminRouter } from "../catalog/routes/index.ts";
+import catalogAdminRouter from "./routes/catalog.ts";
 import { submissionAdminRouter } from "../submission/routes/index.ts";
 import { queryAdminRouter } from "../query/routes/index.ts";
 import { contestAdminRouter } from "../contest/routes/index.ts";
@@ -25,13 +26,14 @@ const router = new Hono<AuthEnv>();
 
 const FINE_GRAINED_ADMIN_PREFIXES = [
   "/api/v1/admin/announcements",
-  "/api/v1/admin/trainings",
-  "/api/v1/admin/problems/review",
 ] as const;
 
 router.use("*", authMiddleware, async (c, next) => {
   if (
-    FINE_GRAINED_ADMIN_PREFIXES.some((prefix) => c.req.path.startsWith(prefix))
+    FINE_GRAINED_ADMIN_PREFIXES.some((prefix) =>
+      c.req.path.startsWith(prefix)
+    ) ||
+    c.req.path.startsWith("/api/v1/admin/catalog")
   ) {
     return next();
   }
@@ -39,7 +41,7 @@ router.use("*", authMiddleware, async (c, next) => {
 });
 
 router.route("/identity", identityAdminRouter);
-router.route("/", catalogAdminRouter);
+router.route("/catalog", catalogAdminRouter);
 router.route("/", submissionAdminRouter);
 router.route("/", queryAdminRouter);
 router.route("/", contestAdminRouter);
