+++ noj-core/src/domains/identity/tests/routes/auth-admin.test.ts
@@ -259,27 +259,31 @@ Deno.test({
 // ─── 题目列表 ────────────────────────────────────────────
 
 Deno.test({
-  name: "admin route: GET /api/v1/admin/problems 非管理员返回 403",
+  name: "admin route: GET /api/v1/admin/catalog/problems 非管理员返回 403",
   ignore: skip,
   sanitizeResources: false,
   sanitizeOps: false,
   fn: async () => {
     const app = createApp();
     const token = await createUserToken();
-    const res = await jsonRequest(app, "/api/v1/admin/problems", { token });
+    const res = await jsonRequest(app, "/api/v1/admin/catalog/problems", {
+      token,
+    });
     assertEquals(res.status, 403);
   },
 });
 
 Deno.test({
-  name: "admin route: GET /api/v1/admin/problems 管理员可访问",
+  name: "admin route: GET /api/v1/admin/catalog/problems 管理员可访问",
   ignore: skip,
   sanitizeResources: false,
   sanitizeOps: false,
   fn: async () => {
     const app = createApp();
     const token = await createUserToken("admin");
-    const res = await jsonRequest(app, "/api/v1/admin/problems", { token });
+    const res = await jsonRequest(app, "/api/v1/admin/catalog/problems", {
+      token,
+    });
     assertEquals(res.status, 200);
     const body = await res.json();
     assertEquals(Array.isArray(body.data), true);
