+++ noj-core/src/shared/db/schema/system.ts
@@ -163,6 +163,9 @@ export const auditLogs = pgTable(
         'problems.delete',
         'problems.runtime_config_changed',
         'problems.imported',
+        'problems.review',
+        'trainings.update',
+        'trainings.delete',
         'tags.create',
         'tags.update',
         'tags.delete',
